django-auth/requirements.txt
django==2.1.1
djangorestframework==3.8.2
django-filter==2.0.0
django-cors-headers==2.4.0
djangorestframework-jwt==1.11.0
markdown==2.6.11